6. Authentication with GraphQL | 15. Locking Down Mutations (Users)

+++ 03-graphql-prisma-authentication/api/src/resolvers/Mutation.js
@@ -55,22 +55,26 @@ const Mutation = {
       )
     };
   },
-  updateUser(parent, args, { prisma }, info) {
+  updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.updateUser(
       {
         where: {
-          id: args.id
+          id: userId
         },
         data: args.data
       },
       info
     );
   },
-  deleteUser(parent, args, { prisma }, info) {
+  deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.deleteUser(
       {
         where: {
-          id: args.id
+          id: userId
         }
       },
       info
@@ -106,7 +110,20 @@ const Mutation = {
       info
     );
   },
-  deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!postExists) {
+      throw new Error('Post not found for this user!');
+    }
+
     return prisma.mutation.deletePost(
       {
         where: {
